refactor(earn): move zapEarn service to earn module
