+++ apps/kyberswap-interface/src/components/ZapCreatePool/CreatePoolModal.tsx
@@ -5,8 +5,8 @@ import { Trans, t } from '@lingui/macro'
 import Portal from '@reach/portal'
 import { useQuery } from '@tanstack/react-query'
 import { useCallback, useEffect, useMemo, useState } from 'react'
+import { useLazyPoolDetailQuery, useSupportedProtocolsQuery } from 'services/earn'
 import { useCheckPairQuery } from 'services/marketOverview'
-import { useLazyPoolDetailQuery, useSupportedProtocolsQuery } from 'services/zapEarn'
 
 import { ReactComponent as DropdownSVG } from 'assets/svg/down.svg'
 import { ButtonOutlined, ButtonPrimary } from 'components/Button'
