+++ apps/kyberswap-interface/src/pages/Earns/PoolExplorer/TableHeader.tsx
@@ -1,6 +1,6 @@
 import { t } from '@lingui/macro'
 import { useMedia } from 'react-use'
-import { PoolQueryParams } from 'services/zapEarn'
+import { PoolQueryParams } from 'services/earn'
 
 import { SortBy } from 'pages/Earns/PoolExplorer'
 import {
