+++ apps/kyberswap-interface/src/pages/Earns/PositionDetail/AnalyticTab.tsx
@@ -1,7 +1,7 @@
 import { ChainId } from '@kyberswap/ks-sdk-core'
 import { t } from '@lingui/macro'
 import { useMemo } from 'react'
-import { PositionHistory, usePositionHistoryQuery } from 'services/zapEarn'
+import { PositionHistory, usePositionHistoryQuery } from 'services/earn'
 
 import CopyHelper from 'components/Copy'
 import TokenLogo from 'components/TokenLogo'
