+++ apps/kyberswap-interface/src/pages/Earns/SmartExitOrders/useSmartExitOrdersData.ts
@@ -1,7 +1,7 @@
 import { t } from '@lingui/macro'
 import { useCallback, useEffect, useMemo, useRef, useState } from 'react'
+import { useUserPositionsQuery } from 'services/earn'
 import { useGetSmartExitOrdersQuery } from 'services/smartExit'
-import { useUserPositionsQuery } from 'services/zapEarn'
 
 import { SmartExitDexType } from 'pages/Earns/components/SmartExit/constants'
 import { EARN_DEXES, Exchange } from 'pages/Earns/constants'
